# src/main/resources/openapi/resolver.json
{
	"openapi": "3.0.2",
	"info":
	{
		"title": "\u041a\u0430\u043c\u0435\u0440\u0430 \u043a\u043e\u043d\u0442\u0440\u043e\u043b\u044f \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438",
		"description": "REST-\u0441\u0435\u0440\u0432\u0438\u0441 \u0432 \u043e\u0441\u043d\u043e\u0432\u043d\u043e\u043c \u043f\u0440\u0435\u0434\u043d\u0430\u0437\u043d\u0430\u0447\u0435\u043d \u0434\u043b\u044f \u043f\u0440\u043e\u0441\u043c\u043e\u0442\u0440\u0430 \u0441\u0432\u0435\u0434\u0435\u043d\u0438\u0439 \u043e \u0437\u0430\u0440\u0435\u0433\u0438\u0441\u0442\u0440\u0438\u0440\u043e\u0432\u0430\u043d\u043d\u044b\u0445 \u0441\u043e\u0431\u044b\u0442\u0438\u044f\u0445 \u043a\u0430\u043c\u0435\u0440\u043e\u0439 \u043a\u043e\u043d\u0442\u0440\u043e\u043b\u044f \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438, \u0430 \u0442\u0430\u043a \u0436\u0435 \u0434\u043b\u044f \u0438\u0437\u043c\u0435\u043d\u0435\u043d\u0438\u044f \u0442\u0435\u043a\u0443\u0449\u0435\u0433\u043e \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u044f \u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u043e\u0439 \u043f\u0440\u0435\u0434\u0435\u043b\u044c\u043d\u043e\u0439 \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438. \u0422\u0430\u043a \u0436\u0435 \u0441\u0435\u0440\u0432\u0438\u0441 \u043f\u043e\u0437\u0432\u043e\u043b\u044f\u0435\u0442 \u0443\u0437\u043d\u0430\u0442\u044c \u0442\u0435\u043a\u0443\u0449\u0435\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u0435 \u043f\u0440\u0435\u0434\u0435\u043b\u044c\u043d\u043e \u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u043e\u0439 \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438.",
		"version": "1.0.0",
		"contact":
		{
			"name": "Pavel Kalinchuk",
			"email": "spr1tee1337@gmail.com"
		}
	},
	"servers": [
	{
		"url": "http://localhost:42099/",
		"description": "test-server"
	}],
	"paths":
	{
		"/camera":
		{
			"post":
			{
				"tags": ["Camera"],
				"summary": "\u0423\u0441\u0442\u0430\u043d\u043e\u0432\u043a\u0430 \u043f\u0440\u0435\u0434\u0435\u043b\u0430 \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438",
				"description": "\u041a\u043e\u043d\u0442\u0440\u0430\u043a\u0442 \u043f\u043e\u0434\u0440\u0430\u0437\u0443\u043c\u0435\u0432\u0430\u0435\u0442 \u0438\u0437\u043c\u0435\u043d\u0435\u043d\u0438\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u044f \u043f\u0440\u0435\u0434\u0435\u043b\u044c\u043d\u043e \u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u043e\u0439 \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438",
				"parameters": [
				{
					"name": "val",
					"description": "\u041d\u043e\u0432\u043e\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u0435 \u043f\u0440\u0435\u0434\u0435\u043b\u044c\u043d\u043e \u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u043e\u0439 \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438",
					"in": "query",
					"required": true,
					"allowEmptyValue": false,
					"schema":
					{
						"type": "integer"
					}
				}],
				"responses":
				{
					"200":
					{
						"description": "Successfull. The current speed limit is 80"
					},
					"400":
					{
						"description": "Bad request"
					}
				}
			}
		},
		"/camera/speed":
		{
			"get":
			{
				"tags": ["Camera"],
				"summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0442\u0435\u043a\u0443\u0449\u0435\u0433\u043e \u043b\u0438\u043c\u0438\u0442\u0430 \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438",
				"description": "\u041f\u043e\u0437\u0432\u043e\u043b\u044f\u0435\u0442 \u043f\u043e\u043b\u0443\u0447\u0438\u0442\u044c \u0442\u0435\u043a\u0443\u0449\u0443\u044e \u043f\u0440\u0435\u0434\u0435\u043b\u044c\u043d\u043e \u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u0443\u044e \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u044c, \u043a\u043e\u0442\u043e\u0440\u0430\u044f \u0443\u0441\u0442\u0430\u043d\u043e\u0432\u043b\u0435\u043d\u0430 \u0432 \u0441\u0435\u0440\u0432\u0438\u0441\u0435",
				"responses":
				{
					"200":
					{
						"description": "\u0412 \u043e\u0442\u0432\u0435\u0442\u0435 \u0431\u0443\u0434\u0435\u0442 \u0443\u043a\u0430\u0437\u0430\u043d\u0430 \u0442\u0435\u043a\u0443\u0449\u0430\u044f \u043f\u0440\u0435\u0434\u0435\u043b\u044c\u043d\u043e \u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u0430\u044f \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u044c. \u041d\u0430\u043f\u0440\u0438\u043c\u0435\u0440, \"80\""
					},
					"400":
					{
						"description": "Bad request"
					}
				}
			}
		},
		"/camera/events":
		{
			"get":
			{
				"tags": ["Camera"],
				"summary": "\u041f\u043e\u043b\u0443\u0447\u0435\u043d\u0438\u0435 \u0437\u0430\u043f\u0438\u0441\u0430\u043d\u043d\u044b\u0445 \u0441\u043e\u0431\u044b\u0442\u0438\u0439",
				"description": "\u041a\u043e\u043d\u0442\u0440\u0430\u043a\u0442 \u0432\u043e\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442 \u043e\u043f\u0440\u0435\u0434\u0435\u043b\u0435\u043d\u043d\u043e\u0435 \u043a\u043e\u043b\u0438\u0447\u0435\u0441\u0442\u0432\u043e \u0437\u0430\u043f\u0438\u0441\u0430\u043d\u043d\u044b\u0445 \u0441\u043e\u0431\u044b\u0442\u0438\u0439. \u0417\u043d\u0430\u0447\u0435\u043d\u0438\u0435 \u0432\u043e\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u043c\u044b\u0445 \u0441\u043e\u0431\u044b\u0442\u0438\u0439 \u043f\u043e \u0443\u043c\u043e\u043b\u0447\u0430\u043d\u0438\u044e \u0440\u0430\u0432\u043d\u043e 10.",
				"parameters": [
				{
					"name": "val",
					"description": "\u041a\u043e\u043b\u0438\u0447\u0435\u0441\u0442\u0432\u043e \u0432\u044b\u0432\u043e\u0434\u0438\u043c\u044b\u0445 \u0437\u0430\u043f\u0438\u0441\u0435\u0439. \u0414\u043e\u043b\u0436\u043d\u043e \u0431\u044b\u0442\u044c \u0431\u043e\u043b\u044c\u0448\u0435 0.",
					"required": false,
					"in": "query",
					"example": "5",
					"schema":
					{
						"type": "integer"
					}
				}],
				"responses":
				{
					"200":
					{
						"description": "\u0423\u0441\u043f\u0435\u0448\u043d\u043e\u0435 \u0432\u044b\u043f\u043e\u043b\u043d\u0435\u043d\u0438\u0435. \u041e\u0442\u0432\u0435\u0442 - \u0441\u043f\u0438\u0441\u043e\u043a \u0441\u043e\u0431\u044b\u0442\u0438\u0439.",
						"content":
						{
							"application/json":
							{
								"schema":
								{
									"title": "Sample",
									"type": "array",
									"items":
									{
										"type": "object",
										"properties":
										{
											"address":
											{
												"type": "string",
												"description": "\u0410\u0434\u0440\u0435\u0441, \u043f\u043e \u043a\u043e\u0442\u043e\u0440\u043e\u043c\u0443 \u0443\u0441\u0442\u0430\u043d\u043e\u0432\u043b\u0435\u043d\u0430 \u043a\u0430\u043c\u0435\u0440\u0430",
												"example": "Russia, Ryazan region, Ryazan, Moscow highway, 21A"
											},
											"speed":
											{
												"type": "integer",
												"description": "\u0421\u043a\u043e\u0440\u043e\u0441\u0442\u044c \u0437\u0430\u0444\u0438\u043a\u0441\u0438\u0440\u043e\u0432\u0430\u043d\u043d\u043e\u0433\u043e \u0441\u043e\u0431\u044b\u0442\u0438\u044f",
												"example": "54"
											},
											"number":
											{
												"type": "string",
												"description": "\u0413\u043e\u0441. \u043d\u043e\u043c\u0435\u0440 \u0430\u0432\u0442\u043e\u043c\u043e\u0431\u0438\u043b\u044f \u0438\u0437 \u0441\u043e\u0431\u044b\u0442\u0438\u044f",
												"example": "\u0410123\u0410\u041062"
											},
											"timestamp":
											{
												"type": "integer",
												"description": "\u0412\u0440\u0435\u043c\u0435\u043d\u043d\u0430\u044f \u043c\u0435\u0442\u043a\u0430 \u0437\u0430\u0444\u0438\u043a\u0441\u0438\u0440\u043e\u0432\u0430\u043d\u043d\u043e\u0433\u043e \u0441\u043e\u0431\u044b\u0442\u0438\u044f",
												"example": "81000000"
											},
											"priorityClass":
											{
												"type": "string",
												"description": "\u041a\u043b\u0430\u0441\u0441 \u0430\u0432\u0442\u043e\u043c\u043e\u0431\u0438\u043b\u044f \u0438\u0437 \u0437\u0430\u0444\u0438\u043a\u0441\u0438\u0440\u043e\u0432\u0430\u043d\u043d\u043e\u0433\u043e \u0441\u043e\u0431\u044b\u0442\u0438\u044f",
												"enum": ["STANDARD", "SPECIAL_SERVICES"],
												"example": "STANDARD"
											},
											"currentLimit":
											{
												"type": "integer",
												"description": "\u0422\u0435\u043a\u0443\u0449\u0435\u0435 \u0437\u043d\u0430\u0447\u0435\u043d\u0438\u0435 \u043f\u0440\u0435\u0434\u0435\u043b\u044c\u043d\u043e \u0434\u043e\u043f\u0443\u0441\u0442\u0438\u043c\u043e\u0439 \u0441\u043a\u043e\u0440\u043e\u0441\u0442\u0438",
												"example": "80"
											},
											"guilty":
											{
												"type": "boolean",
												"description": "\u041f\u0440\u0438\u0437\u043d\u0430\u043a, \u0443\u043a\u0430\u0437\u044b\u0432\u0430\u044e\u0449\u0438\u0439 \u043d\u0430 \u0432\u0438\u043d\u043e\u0432\u043d\u043e\u0441\u0442\u044c \u0432\u043e\u0434\u0438\u0442\u0435\u043b\u044f",
												"example": false
											}
										}
									}
								}
							}
						}
					}
				}
			}
		}
	},
	"tags": [
	{
		"name": "Camera",
		"description": "\u041e\u043f\u0438\u0441\u044b\u0432\u0430\u0435\u0442 \u0434\u0435\u0439\u0441\u0442\u0432\u0438\u044f, \u0441\u043e\u0432\u0435\u0440\u0448\u0430\u0435\u043c\u044b\u0435 \u0441 \u043a\u0430\u043c\u0435\u0440\u043e\u0439"
	}]
}